requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pypdf==3.17.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
openai==1.3.0  # or use transformers for local LLM
rank-bm25==0.2.2
scikit-learn==1.3.0
pydantic==2.4.2
python-dotenv==1.0.0
pytest==7.4.3
pandas==2.1.3
tabulate==0.9.0